changed to mvvm architecture
